Remove MapsValidator
